Add user bar to application layout
Also rearrange a lot of functionality to use the new helper methods
rather than the traditional instance variable. It's a lot safer and
forms a bit better.

However, this should require a reconstruction of the authentication
requirements. Even if that just means a rearrangement and renaming of
various methods in appcontroller.